tests/wpt/tests/css/css-lists/parsing/counter-set-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Lists: parsing counter-set with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-lists-3/#propdef-counter-set">
<meta name="assert" content="counter-set supports the full grammar '[ <counter-name> <integer>? ]+ | none'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value('counter-set', 'none');
test_valid_value('counter-set', 'chapter', 'chapter 0');
test_valid_value('counter-set', 'section -1');
test_valid_value('counter-set', 'first -1 second third 99', 'first -1 second 0 third 99');
test_valid_value('counter-set', 'section calc(1)');
test_valid_value('counter-set', 'section calc(-2.5)');
test_valid_value('counter-set', 'section calc(10 + (5 * sign(2cqw - 10px)))');
</script>
</body>
</html>
