third_party/blink/web_tests/external/wpt/css/css-lists/parsing/counter-set-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['counter-set'] = "section calc(10 + (5 * sign(2cqw - 10px)))" should set the property value
  assert_equals: serialization should be canonical expected "section calc(10 + (5 * sign(2cqw - 10px)))" but got "section calc(10 + 5 * sign(2cqw - 10px))"
Harness: the test ran to completion.
